+++ src/Compute/Compute.Test/ScenarioTests/VirtualMachineScaleSetTests.cs
@@ -486,5 +486,19 @@ namespace Microsoft.Azure.Commands.Compute.Test.ScenarioTests
         {
             TestRunner.RunTestScript("Test-VirtualMachineScaleSetAddProxyAgentExtension");
         }
+
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        public void TestVirtualMachineScaleSetGalleryApplicationFlags()
+        {
+            TestRunner.RunTestScript("Test-VirtualMachineScaleSetGalleryApplicationFlags");
+        }
+
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        public void TestVirtualMachineScaleSetResiliencyView()
+        {
+            TestRunner.RunTestScript("Test-VirtualMachineScaleSetResiliencyView");
+        }
     }
 }